Combine review slice with restaurant slice
