--- src/store/reviewSlice/index.js
@@ -1,2 +1,0 @@
-export * from './reviewSlice';
-export { default } from './reviewSlice';
